+++ src/util/coef.rs
@@ -1,9 +1,9 @@
 /* Treat byte_string as w-bit integers and return index. */
-pub fn coef(byte_string: &[u8], i: u64, w: u64) -> u64 {
-    let index = ((i * w) / 8) as usize;
+pub fn coef(byte_string: &[u8], i: u16, w: u8) -> u64 {
+    let index = ((i * w as u16) / 8) as usize;
 
     let digits_per_byte = 8 / w;
-    let shift = w as u64 * (!i & (digits_per_byte - 1) as u64);
+    let shift = w as u16 * (!i & (digits_per_byte - 1) as u16);
     let mask: u64 = (1 << w) - 1;
 
     (byte_string[index] as u64 >> shift) & mask
